+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 (IN PROGRESS) 0.3.0b0
 =====================
 - Add support for the `increment` and `decrement` commands [#36](https://github.com/pfreixes/emcache/pull/36)
+- Add support for the `touch` command
 
 0.2.1b0
 =======